fixing check 8

+++ 0x14-bit_manipulation/2-get_bit.c
@@ -9,7 +9,7 @@
  */
 int get_bit(unsigned long int n, unsigned int index)
 {
-	unsigned long int ibit;
+	int ibit;
 	if (index >= sizeof(unsigned long int) * 8)
 	{
 		return (-1);
